requirements.txt
torch
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2
torchviz==0.0.2
absl-py
Babel
brotlipy
cached-property
click
contextlib2
cycler
dukpy
einops
Flask
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
future
graphviz
greenlet
grpcio
h5py
itsdangerous
javascripthon
Jinja2
joblib
jupyter-echarts-pypkg
kiwisolver
lml
macropy3
Markdown
matplotlib
ml-collections
numpy
opencv-python
pandas
Pillow
protobuf
pyecharts
pyecharts-javascripthon
pyecharts-jupyter-installer
pyparsing
pywin32
PyYAML
scikit-learn
scipy
SQLAlchemy
tensorboard
tensorboardX
thop
threadpoolctl
timm
torch
torchaudio
torchsummary
torchvision
torchviz
tqdm
ttach
webencodings
Werkzeug
wincertstore
WTForms